+++ internal/util/filesystem_test.go
@@ -26,6 +26,7 @@
 package util
 
 import (
+	"io"
 	"reflect"
 	"testing"
 )
@@ -66,6 +67,68 @@ func TestGetJSONFilenames(t *testing.T) {
 			}
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("GetJSONFilenames() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestExpandPath(t *testing.T) {
+	type args struct {
+		path string
+	}
+	tests := []struct {
+		name            string
+		userHomeDirFunc func() (string, error)
+		args            args
+		wantOut         string
+		wantErr         bool
+	}{
+		{
+			name: "fail",
+			userHomeDirFunc: func() (string, error) {
+				return "", io.EOF
+			},
+			args: args{
+				path: "~",
+			},
+			wantErr: true,
+		},
+		{
+			name: "happy path with home",
+			userHomeDirFunc: func() (string, error) {
+				return "/home/test", nil
+			},
+			args: args{
+				path: "~/test",
+			},
+			wantOut: "/home/test/test",
+		},
+		{
+			name: "happy path relative",
+			userHomeDirFunc: func() (string, error) {
+				return "/home/test", nil
+			},
+			args: args{
+				path: "test",
+			},
+			wantOut: "test",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			old := userHomeDirFunc
+			userHomeDirFunc = tt.userHomeDirFunc
+			defer func() {
+				userHomeDirFunc = old
+			}()
+
+			gotOut, err := ExpandPath(tt.args.path)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ExpandPath() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if gotOut != tt.wantOut {
+				t.Errorf("ExpandPath() = %v, want %v", gotOut, tt.wantOut)
 			}
 		})
 	}
